# StudyCards.UI/studycardsui/openapidoc.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "StudyCards.Server | v1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:5102"
        }
    ],
    "paths": {
...
        "/api/Card/getcards": {
            "get": {
                "tags": [
                    "Card"
                ],
                "parameters": [
                    {
                        "name": "deckId",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CardResponse"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CardResponse"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CardResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Card/updatecard": {
            "put": {
                "tags": [
                    "Card"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCardRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCardRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCardRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Card/addcard": {
            "post": {
                "tags": [
                    "Card"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddCardRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddCardRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddCardRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Card/removecard/deck/{deckId}/card/{cardId}": {
            "delete": {
                "tags": [
                    "Card"
                ],
                "operationId": "removecard",
                "parameters": [
                    {
                        "name": "deckId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cardId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "boolean"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddCardRequest": {
                "type": "object",
                "properties": {
                    "deckId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "cardFront": {
                        "type": "string"
                    },
                    "cardBack": {
                        "type": "string"
                    }
                }
            },
...
            "CardResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "deckId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "cardFront": {
                        "type": "string"
                    },
                    "cardBack": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "UpdateCardRequest": {
                "type": "object",
                "properties": {
                    "cardId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "deckId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "cardFront": {
                        "type": "string"
                    },
                    "cardBack": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
...
}